pyproject.toml
[project]
name = "kasabuttons"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "swilcox", email = "steven@wilcoxzone.com" }]
dependencies = [
    "python-kasa>=0.7.0.dev1",
    "pynput>=1.7.7",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.2.1",
    "loguru>=0.7.2",
    "loki-logger-handler>=0.1.4",
    "requests>=2.32.3",
    "loguru-config>=0.1.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
kasa-buttons = "kasabuttons.kasa_cli:main"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.24.0",
    "pytest>=8.2.1",
    "pytest-asyncio>=0.23.7",
    "asyncmock>=0.4.2",
    "pytest-cov>=5.0.0",
]

[tool.ruff.lint]
select = ["F", "E", "W", "I001"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kasabuttons"]
